+++ ExilenceClient/src/app/authorize/components/player-list/player-list.component.ts
@@ -1,9 +1,10 @@
-import { Component, Input, OnDestroy, OnInit, ViewChild } from '@angular/core';
+import { Component, Input, OnDestroy, OnInit, ViewChild, ChangeDetectorRef } from '@angular/core';
 import { ContextMenuComponent } from 'ngx-contextmenu';
 import { Subscription } from 'rxjs/internal/Subscription';
 
 import { LeagueWithPlayers } from '../../../shared/interfaces/league.interface';
 import { Player } from '../../../shared/interfaces/player.interface';
+import { AccountService } from '../../../shared/providers/account.service';
 import { PartyService } from '../../../shared/providers/party.service';
 
 @Component({
@@ -18,22 +19,29 @@ export class PlayerListComponent implements OnInit, OnDestroy {
 
   private playerLeaguesSub: Subscription;
   private genericPlayersSub: Subscription;
+  private currentPlayerSub: Subscription;
 
-  constructor(public partyService: PartyService) { }
+  constructor(public partyService: PartyService, private accountService: AccountService, private ref: ChangeDetectorRef) { }
 
   playerLeagues: LeagueWithPlayers[];
-
+  currentPlayer: Player;
   genericPlayers: Player[];
 
   ngOnInit() {
+    this.currentPlayerSub = this.accountService.player.subscribe(res => {
+      this.currentPlayer = res;
+    });
     if (!this.localList) {
       this.playerLeaguesSub = this.partyService.playerLeagues.subscribe(res => {
         this.playerLeagues = res;
+        this.ref.detectChanges();
       });
     } else {
       this.genericPlayersSub = this.partyService.genericPlayers.subscribe(res => {
         if (res !== undefined) {
           this.genericPlayers = res;
+          this.ref.detectChanges();
+
           if (this.partyService.selectedGenericPlayerObj === undefined && this.genericPlayers.length > 0) {
             this.partyService.selectedGenericPlayer.next(this.genericPlayers[0]);
           }
@@ -43,12 +51,23 @@ export class PlayerListComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy() {
+    if (this.currentPlayerSub !== undefined) {
+      this.currentPlayerSub.unsubscribe();
+    }
     if (this.playerLeaguesSub !== undefined) {
       this.playerLeaguesSub.unsubscribe();
     }
     if (this.genericPlayersSub !== undefined) {
       this.genericPlayersSub.unsubscribe();
     }
+  }
+
+  assignLeader(characterName: string) {
+    this.partyService.assignLeader(characterName);
+  }
+
+  kickFromGroup(characterName: string) {
+    this.partyService.kickFromParty(characterName);
   }
 
   showMessage(message: any) {