+++ CHANGELOG.md
@@ -2,12 +2,27 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.9.14] - 2019-XX-XX
+### Added
 - Added support for pricing corrupted items (uniques and gems)
+- Added new functionality for "leaders"
+    - The leader can right-click on players to kick them from the group
+    - The leader can right-click on players to pass the leader-role
 - Added the option to select which player to see data for in the currency-tab
 - Added export-functionality to currency summary (for both gain and net worth)
+- Added likelihood-check for Client.txt, displays a warning if it contains old data
+- Added an info-text on the currency-graph if no data has been retrieved
+
+### Changed
+- Now supports up to 40 selected stash-tabs, up from 20
 - Now starts pricing every two minutes, down from three
 - Reworked the in-game overlay to work with the new currency-tab
+
+### Fixed
+- Improved the way we validate session id (no longer requires a Standard-character)
 - Improved the update-functionality for the in-game overlay
+- Data on the player-badges (current zone etc) is now updated instantly
+
+### Removed
 - Removed the smaller variant of the currency-tab
 
 ## [0.9.13] - 2019-01-21